Updated tests so they use docker containers.  Started on docs.  Broke File functionality into File.purs and Stream.purs
