+++ bower.json
@@ -22,7 +22,8 @@
     "purescript-aff": "^2.0.2",
     "purescript-maps": "^2.0.2",
     "purescript-foreign": "^3.0.1",
-    "purescript-generics-rep": "^4.1.0"
+    "purescript-generics-rep": "^4.1.0",
+    "purescript-datetime": "^2.0.0"
   },
   "devDependencies": {
     "purescript-psci-support": "^2.0.0",
