+++ test/Main.purs
@@ -9,9 +9,8 @@ import Control.Monad.Eff.Exception (EXCEPTION)
 import Data.Enum (fromEnum)
 import ExitCodes (ExitCode(Success))
 import PhantomJS.Phantom (exit, PHANTOMJS)
-import PhantomJS.Stream (PHANTOMJSFS)
+import PhantomJS.File (PHANTOMJSFS)
 import Prelude (($), bind)
-import PhantomJS.Stream (PHANTOMJSFS)
 import Test.PhantomJS.Page (pageTests)
 import Test.PhantomJS.Phantom (phantomTests)
 import Test.Unit (describe, it)
